% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_bookmark.R
\name{rtf_bookmark}
\alias{rtf_bookmark}
\title{Insert an RTF bookmark}
\usage{
rtf_bookmark(rtf, bookmark, inc_title = FALSE)
}
\arguments{
\item{rtf}{list. List with 3 elements created by \code{r2rtf::rtf_encode}.}

\item{bookmark}{character. Name of the bookmark to insert. This should be a unique name
that does not contain spaces.}

\item{inc_title}{logical. If TRUE, the bookmark will include the table or figure title. 
Otherwise, if FALSE, the title will omitted. Default is FALSE.}
}
\value{
The input RTF object with the modified start and end elements
}
\description{
\code{rtf_bookmark} inserts an RTF bookmark around the RTF body contents.
}
\details{
\code{rtf_bookmark} modifies the RTF to 
  insert bookmark start and end tags within the document. The bookmark start tag 
  is inserted either at the end of the RTF header (if \code{inc_title} is TRUE) 
  or immediately before the table/figure content in the body (if \code{inc_title} 
  is FALSE). The bookmark end tag is inserted at the beginning of the RTF end element.
}
\examples{
# Create a simple RTF object
rtf <- list(
 start = "{\\\\rtf1\\\\ansi\\\\deff0{\\\\fonttbl{\\\\f0 Arial;}}",
 body = "{\\\\f0\\\\b This is the body of the RTF document.}",
 end = "}"
 )
 
# Insert the bookmark and capture all contents
rtf_bookmark(rtf       = rtf, 
             bookmark  = "mybookmark", 
             inc_title = TRUE)

}
