% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_insert_heading.R
\name{rtf_insert_heading}
\alias{rtf_insert_heading}
\title{Transform RTF Title to Heading}
\usage{
rtf_insert_heading(rtf, heading_level, caption, font.family)
}
\arguments{
\item{rtf}{list. List with 3 elements created by \code{r2rtf::rtf_encode}.}

\item{heading_level}{integer. The heading level to apply to the title (1, 2, 3, or 4).}

\item{caption}{character. The title already present in the RTF.}

\item{font.family}{integer. The font family number used in the RTF.}
}
\value{
The input RTF object with the modified body
}
\description{
\code{rtf_insert_heading} transforms the RTF title into a specified heading style.
}
\details{
\code{rtf_insert_heading} modifies the RTF body to change the title
  into the specified heading style. Note that a stylesheet with defined
  heading level must be present and the title must be bold.
}
