% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_tabstop.R
\name{rtf_tabstop}
\alias{rtf_tabstop}
\title{Insert a right-aligned tab stop in the first line of an RTF page header or footer}
\usage{
rtf_tabstop(rtf, twips, type = "header")
}
\arguments{
\item{rtf}{list. List with 3 elements created by \code{r2rtf::rtf_encode}.}

\item{twips}{integer. The position of the tab stop in twips (1440 twips per inch).}

\item{type}{character. Either "header" or "footer", indicating whether to modify
the page header or footer.}
}
\value{
The input RTF object with the modified body
}
\description{
\code{rtf_tabstop} inserts a tabstop into the first line of an RTF page header or footer.
}
\details{
\code{rtf_tabstop} is an internal function that modifies the RTF body to 
  insert a right-aligned tabstop. It is assumed that the first line of the header 
  or footer already contains a tab character, as is expected, i.e., the first line of the 
  header should be the name of the organization followed by a tab character and the page 
  number field. The first line of the footer should be the data (snapshot) information 
  followed by a tab character and the date of creation.
}
\examples{
# Create a simple RTF object with a header that includes a tab stop
rtf <- list(
  start = "{\\\\rtf1\\\\ansi\\\\deff0{\\\\fonttbl{\\\\f0 Arial;}}",
  body = "{\\\\header\n{\\\\pard{Header Left \\\\tab Header Right}\\\\par}\n}{\\\\f0\\\\b RTF text.}",
  end = "}"
  )
 
# Insert the tabstop
rtf <- rtf_tabstop(rtf, 
                   twips = 9360, 
                   type = "header")

}
