% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf_insert_stylesheet_headings.R
\name{rtf_insert_stylesheet_headings}
\alias{rtf_insert_stylesheet_headings}
\title{Insert Stylesheet for Heading into an RTF file}
\usage{
rtf_insert_stylesheet_headings(rtf)
}
\arguments{
\item{rtf}{list. List with 3 elements created by \code{r2rtf::rtf_encode}. The list 
must have elements named start, body, and end.}
}
\value{
The input RTF object with the stylesheet added.
}
\description{
\code{rtf_insert_stylesheet_headings} inserts a stylesheet defining heading   
  levels 1 through 4 into an RTF file created by \code{r2rtf::rtf_encode}.
}
\details{
\code{rtf_insert_stylesheet_headings} inserts into an RTF a stylesheet that 
  defines heading levels.
}
\examples{
# Create a simple RTF object
rtf <- list(
 start = "{\\\\rtf1\\\\ansi\\\\deff0{\\\\fonttbl{\\\\f0 Arial;}}",
 body = "{\\\\f0\\\\b This is the body of the RTF document.}",
 end = "}"
 )
 
# Insert the stylesheet for headings
rtf_insert_stylesheet_headings(rtf)

}
